move getResourceLabel from new to form component and set props on used pages
